Tests/LibWeb/Text/expected/hit_testing/scrollbar-hit-after-display-list-invalidation.txt

scrolled after scrollbar click: true
